refactor(aceditor):basic split in different files
